LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/orthogonal-wm-container-query-expected.txt

XX

FAIL Initial non-orthogonal width assert_equals: expected 400 but got 50
PASS Orthogonal width
